colour swap

+++ site/htdocs/codex/showItemDetail.php
@@ -86,7 +86,12 @@ echo buildBreadCrumb('codex/items/'.$cleanURL,'The Codex/Item Index/'.$shortname
 	
 
 echo '<h3>'.$shortname.'</h3>';
+
+if(file_exists($_SERVER['DOCUMENT_ROOT'].'/images/game-world/inventory-items/'.$itemID.'.png')) {
 echo '<img src="/images/game-world/inventory-items/'.$itemID.'.png" style="width: auto;" alt="'.$shortname.'">';
+} else {
+echo '<img src="/images/game-world/items/'.$cleanURL.'.png" style="width: auto;" alt="'.$shortname.'">';
+}
 echo '<p>'.str_replace("##contains##", "", $description).'</p>';
